inql.actions.browser: fix a racecondition
when the interceptor is on and you right click on
some query in inql burp, a request is fired to see if GraphIQL
console is reachable, in case the answer is yes, if the proxy is on
and intercepting burp will stop working.

Remove burp proxy from these requests only.