Multiple fixes (#184)
* Removing unused functions

* Quickfix batch queries

* Remove numbers from headings in scanner settings

* Clean up Scanner View

- Moved descriptions to tooltips
- Simplified heading text

* Add visual URL validation when using schema file

Highlights the `urlField` in red and sets a tooltip if the user selects a schema file without specifying a target URL.

* Improve syntax highlight color scheme

- Better colors for light mode
- Small improvements for dark mode

* Fix buttons accent color

* Fix Url and Send button layout in Attacker view

* Hidden profile elements UI from ScanConfigView

* Add cache size limit and configurable setting

Now cache is managed by a memory-aware LRU cache. A config is present to tweak the cache size but it requires to unload and reload the extension to be applied.

* Improve cache management for Formatter

Removed unified caches in globalCache, styles are computed alongside with the formatted text so there is no need to separate the two

* Minor fixes to cache settings GUI

Also removed some debug statements in Formatter.kt

* Add conditional logging to ErrorDialog

* Improved error handling for schema parsing

Error handling when parsing a schema was confusing and would not report the reason for failure.
Now logical errors in a syntactically and semantically correct schema file are reported in the Error tab.

* Improve URL validation

Upon input of invalid URL the urlField borders will turn red and notify the user.

* Fix URL handling

Add todo in Introspection.kt#35. sendRequest() can throw an UnknownHostException inside internal Burp thread and cannot be caught before it unwraps on error logs.

* Fix setLevel

fix mapping to apply values from the SettingsWindow correctly

* Dynamically update config with hooks

When settings are changed using the SettingsWindow a hook is triggered to dynamically apply changes without the need to reload the extension.

For now the only hooks implemented covers
- logging.level
- editor.formatting.cache_size_kb

* Add support for dynamic POI config modification.

- Add option to restart scan of POI and refresh UI accordingly
- Add hooks for
  - proxy.highlight_enabled
  - report.poi.*
- Add settings toggle for custom keywords

* Code cleanup

Applied general code cleanup

* add browser --no-sandbox flag if running as root

If the user is running as root and spawn the embedded browser now the --no-sandbox flag is passed as an argument to prevent the browser from crashing.
solves #177

* implement lazy loading for cycles and POI

* made lazy loading non blocking

* remove useless line in POI text

* make cycle detection a leaf in UI

* Removing comments stripping

* Enabling subscriptions

* Small typos

* Small typos

---------

Co-authored-by: Giacomo Coluccelli <giacomo@doyensec.com>